requirements.txt
blinker==1.4
click==6.7
Flask==1.0.2
Flask-Mail==0.9.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.8.2
Werkzeug==0.14.1
